requirements.txt
streamlit
pymongo
faiss-cpu
sentence-transformers
google-genai
transformers
googletrans==4.0.0-rc1
SpeechRecognition